Rewrite `1pass-search` so it searches through all vault items
This completely changes the behavior of the function.

Previously, your 1password items would have to be titled after the hostname you were connecting to,
with the secret/password stored in a field labeled with your username.

Now it will search through all vault items for any that have a URL value -- as specified with the
`--url` flag of the `op item edit` command -- which matches the host your looking for.

Furthermore, it will then search that subset of items to find one which contains a `username` field
matching the `user` specified in the `SPEC` passed to the search function.

It will either return the first item with a matching username, or the first item returned for the
requested `HOST`.